+++ src/back/template_instantiator_derived.cpp
@@ -24,10 +24,10 @@ void StructTemplateInstantiator::applyNode( TemplateNode& node )
 	{
 		case NODE_TYPE::FOR_EACH_OF_MEMBERS:
 		{
-			unsigned int memberCnt = structure->getChildCount();
-			for ( unsigned int j=0; j<memberCnt; j++ )
+			size_t memberCnt = structure->getChildCount();
+			for ( size_t j=0; j<memberCnt; j++ )
 			{
-				for ( unsigned int k=0; k<node.childNodes.size(); k++ )
+				for ( size_t k=0; k<node.childNodes.size(); k++ )
 				{
 					BackDataMember* member = dynamic_cast<BackDataMember*>( structure->getMember( j ) );
 					if ( member != NULL )
